requirements.txt
# Base dependencies for BioNeuralNet

# PyTorch and related libraries
torch==2.0.0
torch_geometric

# Data handling and processing
PyYAML>=5.4
pandas>=1.0
numpy>=1.18
scikit-learn>=0.24

# Graph and network analysis
node2vec>=0.4.1
networkx

# Visualization and monitoring
matplotlib
ray[tune]
tensorboardX
